update compile_mac.sh to only rebuilds what changed update oled_time.c to move text from left to right

+++ compile_mac.sh
@@ -3,17 +3,30 @@ set -e
 
 echo "Starting tasks..."
 
-cp oled_time.c main.c
-rm -rf build
-mkdir build
-cd build
+# Optional clean build:
+# ./compile_mac.sh clean
+if [ "$1" = "clean" ]; then
+    echo "Removing build folder..."
+    rm -rf build
+fi
 
-cmake -G Ninja ..
-ninja
+cp oled_time.c main.c
+
+echo "Configuring project..."
+
+cmake -S . -B build -G Ninja
+
+echo "Building project..."
+
+cmake --build build
+
+echo "Preparing Raspberry Pi 5 target folder..."
+
+ssh raspi 'mkdir -p ~/pico_cpp'
 
 echo "Sending UF2 to Raspberry Pi 5..."
 
-scp pico_app.uf2 raspi:~/pico_cpp/pico_app.uf2
+scp build/pico_app.uf2 raspi:~/pico_cpp/pico_app.uf2
 
 echo "Flashing Pico from Raspberry Pi 5..."
